Implemented missing functions to link medic with medical_fields and clinics with medical_studies + updated interfaces + revisions to models + revisions to Dao implementations

+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/ResultDao.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface ResultDao {
 
-    public Optional<Result> findById(long id);
+    public Optional<Result> findById(int id);
     public Collection<Result> findByOrderId(long id);
 
     Result register(long order_id, byte[] result_data, byte[] identification, Date date, String responsible_name, String responsible_licence_number);
